Create downloadtags.sh to get encoded song links
I created an IFTTT recipe that will append any of my own tweets that
countain the #Shazam in them to shazamtags.txt file in dropbox
/Music/shazamsnatch.

I created the downloadtags.sh that will use the shazamtags.txt file to
 create the curllinks.sh file, mainly using sed. The curllinks.sh
file uses a Duck Duck Go "I'm feeling Ducky" search url (created by
the downloadtags.sh file). The downloadtags.sh file then calls the
curllinks.sh file, which follows the search URL and gets a URL for a
video of the song on youtube. The youtube links are encoded and put
into a new file called encodedlinks.txt.